pyproject.toml
[project]
name = "excel2word"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=5.31.0",
    "humanize>=4.12.3",
    "polars[excel]>=1.30.0",
    "pydantic>=2.11.5",
    "python-docx>=1.1.2",
]
